requirements.txt
annotated-types==0.7.0
anyio==4.7.0
beautifulsoup4==4.12.3
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
deep-translator==1.11.4
distro==1.9.0
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.12.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.5
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy~=2.1.3
openai==1.57.0
packaging==24.2
pillow==11.0.0
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rich==13.9.4
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers~=4.50.1
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.2.3

onnxruntime~=1.21.0
model2vec~=0.4.1
onnx~=1.17.0
onnxconverter-common~=1.14.0
fasttext~=0.9.3